Fixed integration

+++ src/app/services/expenses-services/billServices/bill.service.ts
@@ -2,6 +2,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Bill } from '../../../models/expenses-models/bill';
+import { UserService } from '../userServices/user.service';
 
 @Injectable({
   providedIn: 'root'
@@ -11,16 +12,19 @@ export class BillService {
   
   private apiUrl = 'http://localhost:8080/expenses/getByFilters';
 
-  constructor(private http: HttpClient) {}
+  constructor(private http: HttpClient,private userService :UserService) {}
 
  
   deleteLogicBill(id: number): Observable<void> {
-    const url = 'http://localhost:8080/expenses?id='+id;
+    const url = `http://localhost:8080/expenses?id=${id}&userId=${this.userService.getUserId()}`;
     const headers = new HttpHeaders({ 'Content-Type': 'application/json' });
     return this.http.delete<void>(url, { headers });
   }
-  createNoteOfCredit(failedBillId: number | null) {
-    const url = 'http://localhost:8080/expenses/note_credit?id='+failedBillId;
+  createNoteOfCredit(failedBillId: number | null): Observable<void> {
+    if (failedBillId === null) {
+      throw new Error('El ID de la factura fallida no puede ser null.');
+    }
+    const url = `http://localhost:8080/expenses/note_credit?id=${failedBillId}&userId=${this.userService.getUserId()}`;
     const headers = new HttpHeaders({ 'Content-Type': 'application/json' });
     return this.http.delete<void>(url, { headers });
   }
